// src/types/userResumeInputTypes.ts
export interface Experience {
  period: string;
  position: string;
  company: string;
  description: string;
}

export interface Education {
  school: string;
  major: string;
  gpa: string;
  period: string;
  description: string;
}

export interface Skills {
  technical: string;
  nonTechnical: string;
  managerial: string;
  soft: string;
}